write all components, reducers and actions test
