+++ client/__mocks__/router.js
@@ -1,0 +1,10 @@
+/**
+ * @function Router
+ * @return {void} null
+ */
+export default function Router() {
+  this.router = [];
+  this.push = (a) => { this.router.push(a); };
+  this.get = index => this.router[index];
+  this.length = () => this.router.length;
+}
