+++ package.json
@@ -59,6 +59,7 @@
     "node-env-file": "^0.1.8",
     "node-sass": "^4.5.3",
     "nodemailer": "^4.0.1",
+    "nodemon": "^1.12.1",
     "nyc": "^11.1.0",
     "path": "^0.12.7",
     "pg": "^6.4.0",
@@ -109,6 +110,7 @@
     "nightwatch": "^0.9.16",
     "nock": "^9.0.14",
     "react-addons-test-utils": "^15.6.0",
+    "react-mock-router": "^1.0.11",
     "react-test-renderer": "^15.6.1",
     "redux-mock-store": "^1.2.3",
     "sass-loader": "^6.0.6",
@@ -119,6 +121,10 @@
     "uglifyjs-webpack-plugin": "^0.4.6"
   },
   "jest": {
+    "globals": {
+      "widow": true
+    },
+    "verbose": true,
     "testEnvironment": "jsdom",
     "transform": {
       ".*": "<rootDir>/node_modules/babel-jest"
@@ -129,7 +135,8 @@
       "**/components/**",
       "!**/dist/**",
       "!**/js/**",
-      "!**/utils/**"
+      "!**/utils/**",
+      "!**/client/store/*"
     ],
     "moduleFileExtensions": [
       "js"
